+++ src/components/BarChart/examples/05.limited-ticks.jsx
@@ -22,10 +22,14 @@ const data = [
 	{ x: '2015-01-17', y: 6 },
 ];
 
+const style = {
+	paddingTop: '4rem',
+};
+
 export default createClass({
 	render() {
 		return (
-			<div>
+			<div style={style}>
 				<BarChart
 					width={750}
 					data={data}