+++ src/components/Icon/DotsIcon/DotsIcon.tsx
@@ -4,16 +4,31 @@ import Icon, { IIconProps, propTypes as iconPropTypes } from '../Icon';
 import { lucidClassNames } from '../../../util/style-helpers';
 import { omitProps } from '../../../util/component-types';
 import { Color } from '../Icon';
+import PropTypes from 'react-peek/prop-types';
 
 const cx = lucidClassNames.bind('&-DotsIcon');
 
-interface IDotsIconProps extends IIconProps {}
+const { oneOf } = PropTypes;
+interface IDotsIconProps extends IIconProps {
+	direction?: 'vertical' | 'horizontal';
+}
 
 export const DotsIcon = ({
 	className,
+	direction = 'horizontal',
 	color = Color.primary,
 	...passThroughs
 }: IDotsIconProps) => {
+	const isVerticalOrientation = direction === 'vertical';
+	const leftOrTopPosition = {
+		cx: isVerticalOrientation ? '8' : '14.5',
+		cy: isVerticalOrientation ? '14.5' : '8',
+	};
+	const rightOrBottomPosition = {
+		cx: isVerticalOrientation ? '8' : '1.5',
+		cy: isVerticalOrientation ? '1.5' : '8',
+	};
+	
 	return (
 		<Icon
 			{...omitProps(passThroughs, undefined, _.keys(DotsIcon.propTypes), false)}
@@ -22,8 +37,8 @@ export const DotsIcon = ({
 			className={cx('&', className)}
 		>
 			<circle className={cx(`&-color-${color}`)} cx='8' cy='8' r='1' />
-			<circle className={cx(`&-color-${color}`)} cx='14.5' cy='8' r='1' />
-			<circle className={cx(`&-color-${color}`)} cx='1.5' cy='8' r='1' />
+			<circle className={cx(`&-color-${color}`)} {...leftOrTopPosition} r='1' />
+			<circle className={cx(`&-color-${color}`)} {...rightOrBottomPosition} r='1' />
 		</Icon>
 	);
 };
@@ -37,7 +52,17 @@ DotsIcon.peek = {
 	extend: 'Icon',
 	madeFrom: ['Icon'],
 };
-DotsIcon.propTypes = iconPropTypes;
-DotsIcon.defaultProps = Icon.defaultProps;
+
+DotsIcon.propTypes = {
+	...iconPropTypes,
+	direction: oneOf(['vertical', 'horizontal'])`
+		Sets the orientation of how the dots are displayed. Defaults to 'horizontal'. 
+	`,
+};
+
+DotsIcon.defaultProps = {
+	...Icon.defaultProps,
+	direction: 'horizontal',
+};
 
 export default DotsIcon;