+++ src/components/PieChart/examples/02.basic-donuts.tsx
@@ -12,10 +12,14 @@ const data = [
 	{ x: 'Ben', y: 15 },
 ];
 
+const style = {
+	paddingTop: '4rem',
+};
+
 export default createClass({
 	render() {
 		return (
-			<div>
+			<div style={style}>
 				<PieChart data={data} isDonut />
 				<PieChart
 					data={data}